+++ FlOpEDT/TTapp/helpers/minhalfdays.py
@@ -85,7 +85,7 @@ class MinHalfDaysHelperModule(MinHalfDaysHelperBase):
 
             # add constraint linking MBHD to TT
             for apm in [Time.AM, Time.PM]:
-                halfdayslots = set(sl for sl in self.ttmodel.wdb.slots if sl.day == d and sl.apm == apm)
+                halfdayslots = set(sl for sl in self.ttmodel.wdb.courses_slots if sl.day == d and sl.apm == apm)
                 card = len(halfdayslots)
                 expr = self.ttmodel.lin_expr()
                 expr += card * mod_b_h_d[(self.module, d, apm)]
@@ -124,7 +124,7 @@ class MinHalfDaysHelperModule(MinHalfDaysHelperBase):
 class MinHalfDaysHelperGroup(MinHalfDaysHelperBase):
 
     def build_variables(self):
-        courses = self.ttmodel.wdb.courses.filter(group=self.group, week=self.week)
+        courses = self.ttmodel.wdb.courses.filter(groups=self.group, week=self.week)
 
         expression = self.ttmodel.check_and_sum(
             self.ttmodel.GBHD,
@@ -173,13 +173,13 @@ class MinHalfDaysHelperTutor(MinHalfDaysHelperBase):
         if self.constraint.join2courses and len(courses) in [2, 4]:
             for d in days:
                 for c in courses:
-                    sl8h = min(slots_filter(self.ttmodel.wdb.slots, day=d, apm=Time.AM) & self.ttmodel.wdb.compatible_slots[c])
-                    sl14h = min(slots_filter(self.ttmodel.wdb.slots, day=d, apm=Time.PM) & self.ttmodel.wdb.compatible_slots[c])
+                    sl8h = min(slots_filter(self.ttmodel.wdb.courses_slots, day=d, apm=Time.AM) & self.ttmodel.wdb.compatible_slots[c])
+                    sl14h = min(slots_filter(self.ttmodel.wdb.courses_slots, day=d, apm=Time.PM) & self.ttmodel.wdb.compatible_slots[c])
                     for c2 in courses.exclude(id=c.id):
                         sl11h = max(
-                            slots_filter(self.ttmodel.wdb.slots, day=d, apm=Time.AM) & self.ttmodel.wdb.compatible_slots[c2])
+                            slots_filter(self.ttmodel.wdb.courses_slots, day=d, apm=Time.AM) & self.ttmodel.wdb.compatible_slots[c2])
                         sl17h = max(
-                            slots_filter(self.ttmodel.wdb.slots, day=d, apm=Time.PM) & self.ttmodel.wdb.compatible_slots[c2])
+                            slots_filter(self.ttmodel.wdb.courses_slots, day=d, apm=Time.PM) & self.ttmodel.wdb.compatible_slots[c2])
                         if self.constraint.weight:
                             conj_var_AM = self.ttmodel.add_conjunct(self.ttmodel.TT[(sl8h, c)],
                                                                     self.ttmodel.TT[(sl11h, c2)])